# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent API",
    "version": "1.0.0"
  },
...
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/signed-url": {
      "post": {
        "operationId": "getSignedUrl",
        "parameters": [],
        "tags": [
          "files"
        ],
        "description": "Generate a signed URL for uploading files to cloud storage. Creates a file record and returns a time-limited upload URL. Supports PDF, PNG, JPG, and DICOM file types.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "name",
                  "type"
                ],
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "application/pdf",
                      "image/png",
                      "image/jpeg",
                      "image/jpg",
                      "application/dicom"
                    ]
                  },
                  "size": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response for 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "signedUrl": {
                      "type": "string"
                    },
                    "fileId": {
                      "type": "string"
                    },
                    "requiredHeaders": {
                      "type": "object",
                      "properties": {
                        "Content-Type": {
                          "anyOf": [
                            {
                              "enum": [
                                "application/pdf"
                              ],
                              "type": "string"
                            },
                            {
                              "enum": [
                                "image/png"
                              ],
                              "type": "string"
                            },
                            {
                              "enum": [
                                "image/jpeg"
                              ],
                              "type": "string"
                            },
                            {
                              "enum": [
                                "image/jpg"
                              ],
                              "type": "string"
                            },
                            {
                              "enum": [
                                "application/dicom"
                              ],
                              "type": "string"
                            }
                          ]
                        },
                        "x-goog-if-generation-match": {
                          "enum": [
                            "0"
                          ],
                          "type": "string"
                        },
                        "x-goog-meta-file-id": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "Content-Type",
                        "x-goog-if-generation-match",
                        "x-goog-meta-file-id"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "signedUrl",
                    "fileId",
                    "requiredHeaders"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ValidationError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedErr"
                }
              }
            }
          }
        }
      }
    },
    "/files": {
      "get": {
        "operationId": "getFiles",
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "files"
        ],
        "description": "Get a list of files uploaded by a specific user, ordered by creation date. Returns file metadata including name, path, content type, and size.",
        "responses": {
          "200": {
            "description": "Response for 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "records": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "string"
                          },
                          "id": {
                            "type": "string"
                          },
                          "userId": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          },
                          "size": {
                            "anyOf": [
                              {
                                "type": "null"
                              },
                              {
                                "type": "number"
                              }
                            ]
                          },
                          "contentType": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "path",
                          "id",
                          "userId",
                          "name",
                          "updatedAt",
                          "size",
                          "contentType"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "records"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ValidationError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedErr"
                }
              }
            }
          }
        }
      }
    },
    "/files/{id}": {
      "get": {
        "operationId": "getFileById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "tags": [
          "files"
        ],
        "description": "Get detailed information about a specific file by its ID. Returns complete file metadata and properties.",
        "responses": {
          "200": {
            "description": "Response for 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "size": {
                      "anyOf": [
                        {
                          "type": "null"
                        },
                        {
                          "type": "number"
                        }
                      ]
                    },
                    "contentType": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "path",
                    "id",
                    "userId",
                    "name",
                    "updatedAt",
                    "size",
                    "contentType"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ValidationError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedErr"
                }
              }
            }
          }
        }
      }
    },
    "/files/{id}/signed-url": {
      "get": {
        "operationId": "getSignedReadUrl",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "tags": [
          "files"
        ],
        "description": "Generate a short-lived signed URL for reading a file by its ID. Ensures the file belongs to the authenticated user.",
        "responses": {
          "200": {
            "description": "Response for 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "signedUrl": {
                      "type": "string"
                    },
                    "expiresAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "signedUrl",
                    "expiresAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ValidationError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedErr"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UnauthorizedErr": {
        "type": "object",
        "required": [
          "type",
          "title"
        ],
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "Unauthorized"
            ]
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}